feat(context): base context runs the walk, through a seam the hook shares
`base context <text>` claimed in its own help string to be the same engine as
hook injection. It was the same engine for the DOMAIN layer only: the arm had
no walk, so it could not preview a base-context block for anything a user
named. On the operator's store it returned byte-identical output for three
different inputs including a nonsense control -- the argument selected
nothing.

walk::walk_from_text is the new single seam: it builds the maps, builds the
transient and supersede closures, and walks. Both the prompt hook and
context_pull call it, so the two paths cannot drift apart -- and drift in
exactly this seam is what produced N-WALK-DEAD-WITHOUT-A-MATCHED-DOMAIN.

context_pull also loses its `matched.is_empty()` early return. The walk
resolves what the TEXT names, which has nothing to do with whether a domain
trigger fired; returning early made the command dead on any machine without
an always-on domain. The store was already parsed above that point, so
nothing extra is paid to reach the walk.

The served list that query_domain_from_graph already computes is now
collected rather than discarded, so a record the domain block just printed is
not reprinted four lines later under a base-context heading. Records dedup; a
served root still walks and is never listed as its own record.

No session dedup on this path, in either direction: a preview that marked a
walk key would silently suppress the hook's block on the next real prompt.
